Remove package-lock.json for compromising issues
The package-lock.json is a hindrance because when you run a beginning install for projects it isn't compatible as the package-lock.json is created after the initial install. You also do not have node_modules so it in full will cause more harm than help. I suggest removing the package-lock.json and having the end user create their own through either yarn, npm or pnpm or some other node package manager that is suitable for installing dependencies.